Make USB cameras work with Android pie
1. BaseActivity inherits from AppCompatActivity. Changed all the styles to use AppCompat themes. Included appcompat-v7 in build files.
2. Made the build work with proguard by keeping all the USB classes
3. Modified device_filter.xml files to include <<usb-device class="14" subclass="2" />, this makes our camera work for Android pie

+++ libuvccamera/src/main/jni/utilbase.h
@@ -203,7 +203,7 @@
 #endif
 
 #define		ENTER()				LOGD("begin")
-#define		RETURN(code,type)	{type RESULT = code; LOGD("end (%d)", (int)RESULT); return RESULT;}
+#define		RETURN(code,type)	{type RESULT = code; LOGD("end (%d)", (long)RESULT); return RESULT;}
 #define		RET(code)			{LOGD("end"); return code;}
 #define		EXIT()				{LOGD("end"); return;}
 #define		PRE_EXIT()			LOGD("end")
